+++ src/main/java/searchengine/dto/PageDto.java
@@ -1,8 +1,10 @@
 package searchengine.dto;
 
-public record PageDto(Integer id,
-                      Integer siteId,
-                      String path,
-                      Integer code,
-                      String content) {
+public record PageDto(
+        Long id,
+        Long siteId,
+        String path,
+        Integer code,
+        String content
+) {
 }
