+++ src/main/java/searchengine/dto/SiteDto.java
@@ -4,10 +4,12 @@ import searchengine.model.Status;
 
 import java.time.LocalDateTime;
 
-public record SiteDto(Integer id,
-                      Status status,
-                      LocalDateTime statusTime,
-                      String lastError,
-                      String url,
-                      String name) {
+public record SiteDto(
+        Long id,
+        Status status,
+        LocalDateTime statusTime,
+        String lastError,
+        String url,
+        String name
+) {
 }
