+++ src/main/java/searchengine/repository/LemmaRepository.java
@@ -7,7 +7,9 @@ import searchengine.model.Site;
 import java.util.List;
 import java.util.Set;
 
-public interface LemmaRepository extends JpaRepository<Lemma, Integer> {
+public interface LemmaRepository extends JpaRepository<Lemma, Long> {
+
     List<Lemma> findAllBySiteAndLemmaIn(Site site, Set<String> lemmaTexts);
     List<Lemma> findAllByLemmaInAndSiteIn(Set<String> lemmas, List<Site> sites);
+
 }
