+++ src/main/java/searchengine/repository/SiteRepository.java
@@ -7,6 +7,8 @@ import searchengine.model.Site;
 import java.util.Optional;
 
 @Repository
-public interface SiteRepository extends JpaRepository<Site, Integer> {
+public interface SiteRepository extends JpaRepository<Site, Long> {
+
     Optional<Site> findByUrl(String url);
+
 }
